Data: Increased size to uint32_t (prepare for large transfers)
Increased Data size from uint16_t to uint32_t in preparation for
supporting transfers > 64kb. Adjusted transfer methods and callbacks
in Userport and Service accordingly.

Also fixed memory leak in Request introduced in recent refactoring.

+++ wic64/callback.h
@@ -4,7 +4,7 @@
 #include <cstdint>
 
 namespace WiC64 {
-    typedef void (*callback_t) (uint8_t* data, uint16_t size);
+    typedef void (*callback_t) (uint8_t* data, uint32_t size);
 }
 
 #endif //WIC64_CALLBACK_H
